py_backend/requirements.txt
# Web framework
fastapi
uvicorn
python-multipart  # For file uploads
pydantic # Using 1.x for better compatibility

# Database
psycopg2-binary
pgvector
sqlalchemy
alembic

# Authentication
python-jose[cryptography]
passlib[bcrypt]
python-dotenv

# LangChain and related
langchain
langchain-community
langchain-core
chromadb
langchain-chroma
langchain-ollama

# Document processing
pypdf
python-docx
pandas

# Utilities
requests

# IMPORTANT: Keep numpy locked at a compatible version
numpy

faster-whisper

diffusers[torch]
torch
transformers
safetensors
accelerate